[BKR-19] Uninitialized Implementation, Vault using revert instead of requires (#19)
* Proxy

* Does not allow to initialize the implementation vault

* Settings behind a proxy

* Vault

* Strategy , Vault and Settings Deploy only with proxies

* Proxy only deploys

* Settings

* Removed Unused code

* Settings

* Vault Improved

* Service Registry with Custom Errors

* Settings

* Tests layout improved

* Baker

* Vault Proxy updated

* Strategy

* Fixed typescript errors

* Fixes Typescript issues

* Typescript hell

* fix: types and lint (#20)

* test: fix types and lint

* fix: types and lint

---------

+++ contracts/interfaces/core/ISettings.sol
@@ -182,7 +182,7 @@ interface ISettings {
 
     function setMaxDepositInETH(uint256 value) external;
 
-    function setOraclePriceMaxAge(uint value) external;
+    function setOraclePriceMaxAge(uint256 value) external;
 
-    function getOraclePriceMaxAge() external view returns (uint);
+    function getOraclePriceMaxAge() external view returns (uint256);
 }
